+++ docs/en/changes/changes.md
@@ -23,6 +23,7 @@
 * `BrowserErrorLog`, OAP Server generated UUID to replace the original client side ID, because Browser scripts can't guarantee generated IDs are globally unique.
 * MQE: fix multiple labeled metric query and ensure no results are returned if no label value combinations match.
 * Fix `BrowserErrorLog` BanyanDB storage query order.
+* `BanyanDB Client`: Property query support `Order By`.
 
 #### UI
 * Fix the missing icon in new native trace view.